Request to pull "prodAllLanes for WEFC_Toronto" from WEFC_Toronto_reports

+++ fannie/cron/Suspensions/reactivate.equity.php
@@ -62,7 +62,7 @@ $custQ = "UPDATE custdata AS c LEFT JOIN {$TRANS}newBalanceStockToday_test AS s
 $sql->query($custQ);
 
 $histQ = "insert into suspension_history
-	    select 'automatic',".$sql->now().",
+	    select 'equity paid',".$sql->now().",
 	    'Account reactivated',c.CardNo,0 from
 	    suspensions as s left join
 	    custdata as c on s.cardno=c.CardNo
@@ -82,8 +82,10 @@ while($clearW = $sql->fetch_row($clearR)){
 }
 $cns = rtrim($cns,",").")";
 
-$delQ = "DELETE FROM suspensions WHERE cardno IN $cns";
-$delR = $sql->query($delQ);
+if ($cns != "()"){
+	$delQ = "DELETE FROM suspensions WHERE cardno IN $cns";
+	$delR = $sql->query($delQ);
+}
 
 
 ?>
